requirements.txt
sentence-transformers>=2.2.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
mysql-connector-python>=8.0.26
numpy>=1.21.0
pydantic>=1.8.0 